#965: Refactored test
